+++ event-sourcing/event-source/src/main/java/uk/gov/justice/services/eventsourcing/source/core/EventAppender.java
@@ -7,6 +7,8 @@ import static uk.gov.justice.services.messaging.JsonEnvelope.metadataFrom;
 
 import uk.gov.justice.services.eventsourcing.publisher.jms.EventPublisher;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.EventRepository;
+import uk.gov.justice.services.eventsourcing.repository.jdbc.eventstream.EventStream;
+import uk.gov.justice.services.eventsourcing.repository.jdbc.eventstream.EventStreamJdbcRepository;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.exception.StoreEventRequestFailedException;
 import uk.gov.justice.services.eventsourcing.source.core.exception.EventStreamException;
 import uk.gov.justice.services.messaging.JsonEnvelope;
@@ -19,8 +21,12 @@ import javax.inject.Inject;
 @ApplicationScoped
 public class EventAppender {
 
+    private static final long INITIAL_STREAM_EVENT = 1L;
     @Inject
     EventRepository eventRepository;
+
+    @Inject
+    EventStreamJdbcRepository streamRepository;
 
     @Inject
     EventPublisher eventPublisher;
@@ -34,6 +40,9 @@ public class EventAppender {
      */
     void append(final JsonEnvelope event, final UUID streamId, final long version) throws EventStreamException {
         try {
+            if (version == INITIAL_STREAM_EVENT) {
+                streamRepository.insert(new EventStream(streamId));
+            }
             final JsonEnvelope eventWithStreamIdAndVersion = eventFrom(event, streamId, version);
             eventRepository.store(eventWithStreamIdAndVersion);
             eventPublisher.publish(eventWithStreamIdAndVersion);